Convert raw replay data into full osr replay file

+++ src/replays/replay_manager.cc
@@ -3,10 +3,10 @@
         #include <filesystem>
 
         #if defined(_WIN32)
-        // Workaround for Windows: https://docs.microsoft.com/en-us/cpp/standard-library/filesystem
-        namespace fs = std::experimental::filesystem::v1;
+            // Workaround for Windows: https://docs.microsoft.com/en-us/cpp/standard-library/filesystem
+            namespace fs = std::experimental::filesystem::v1;
         #else
-        namespace fs = std::filesystem;
+            namespace fs = std::filesystem;
         #endif
     #else
         #include <experimental/filesystem>
@@ -15,12 +15,21 @@
 #else
     // Compiler doesn't support C++17 __has_include so we can expect filesystem is in experimental state
     #include <experimental/filesystem>
-
-namespace fs = std::experimental::filesystem;
+    namespace fs = std::experimental::filesystem;
 #endif
 
+#include <boost/iostreams/filtering_streambuf.hpp>
+#include <boost/iostreams/filter/zlib.hpp>
+#include <boost/iostreams/filter/lzma.hpp>
+#include <boost/iostreams/copy.hpp>
 #include <fstream>
 
+#include "../io/osu_buffer.hh"
+#include "../users/user_manager.hh"
+#include "../thirdparty/loguru.hh"
+#include "../utils/crypto.hh"
+#include "../utils/osu_string.hh"
+#include "../utils/time_utils.hh"
 #include "replay_manager.hh"
 
 static std::string dir = fs::current_path().u8string() + fs::path::preferred_separator + "replays";
@@ -30,17 +39,93 @@ void shiro::replays::init() {
         fs::create_directories(dir);
 }
 
-void shiro::replays::save_replay(const shiro::scores::score &s, std::string replay) {
+void shiro::replays::save_replay(const shiro::scores::score &s, int32_t game_version, std::string replay) {
     std::string filename = dir + fs::path::preferred_separator + std::to_string(s.id) + ".osr";
+
+    std::shared_ptr<users::user> user = users::manager::get_user_by_id(s.user_id);
+
+    if (user == nullptr)
+        return;
 
     if (fs::exists(filename))
         fs::remove(filename);
 
+    // Build magic hash of replay
+    char hash_buffer[1024];
+
+    std::snprintf(hash_buffer, sizeof(hash_buffer), "%ip%io%io%it%ia%ir%se%iy%so%su%li%s%i%s",
+            s._100_count, s._300_count, s._50_count, s.gekis_count, s.katus_count, s.miss_count,
+            s.beatmap_md5.c_str(), s.max_combo, s.fc ? "True" : "False",
+            user->presence.username.c_str(), s.total_score, s.rank.c_str(), s.mods, "True");
+
+    // Convert raw replay into full osu! replay file
+
+    std::string beatmap_md5 = utils::osu_string(s.beatmap_md5);
+    std::string username = utils::osu_string(user->presence.username);
+    std::string hash = utils::osu_string(utils::crypto::md5::hash(hash_buffer));
+
+    io::buffer buffer;
+
+    buffer.write<uint8_t>(s.play_mode);
+    buffer.write<int32_t>(game_version);
+
+    buffer.write_string(beatmap_md5);
+    buffer.write_string(username);
+    buffer.write_string(hash);
+
+    buffer.write<int16_t>(s._300_count);
+    buffer.write<int16_t>(s._100_count);
+    buffer.write<int16_t>(s._50_count);
+    buffer.write<int16_t>(s.gekis_count);
+    buffer.write<int16_t>(s.katus_count);
+    buffer.write<int16_t>(s.miss_count);
+
+    buffer.write<int32_t>(s.total_score);
+    buffer.write<int16_t>(s.max_combo);
+    buffer.write<uint8_t>(s.fc);
+    buffer.write<int32_t>(s.mods);
+
+    buffer.write_string(utils::osu_string("15033|1,7085|1,9207|1,11394|1,13850|1,15852|1,17853|1,19993|1,22007|1,24162|1,26313|1,28466|1,30505|1,32624|1,34777|1,36922|1,39085|1,41393|1,43545|1,45691|1,47696|0.99,49698|1,52005|1,54168|1,56170|1,58320|1,60387|1,62445|1,64468|1,66773|1,68928|1,71095|0.9,73395|1,75399|1,77856|1,80316|1,82466|1,84620|0.85,86777|0.98,88932|1,90933|1,91995|1,"));
+    buffer.write<int64_t>(utils::time::get_current_time_ticks());
+
+    buffer.write<int32_t>(replay.size());
+    buffer.write_string(replay);
+
+    buffer.write<int64_t>(s.id);
+
     std::ofstream stream(filename, std::ios::trunc);
-    stream << replay;
+    stream << buffer.serialize();
     stream.close();
+
+    // If the replay is bigger than 1 mib, compress it
+    // TODO: Implement a smarter compression method than zlib
+    if (fs::file_size(filename) >= 1048576) {
+        fs::remove(filename);
+
+        filename = dir + fs::path::preferred_separator + std::to_string(s.id) + ".osr.zz";
+
+        std::stringstream original;
+        std::stringstream compressed;
+
+        original << buffer.serialize();
+
+        boost::iostreams::filtering_streambuf<boost::iostreams::input> output;
+        output.push(boost::iostreams::zlib_compressor());
+        output.push(original);
+
+        boost::iostreams::copy(output, compressed);
+
+        stream = std::ofstream(filename, std::ios::trunc);
+        stream << compressed.str();
+        stream.close();
+
+        LOG_S(WARNING) << "Uncompressed replay was >1mb, saved replay with zlib compression.";
+    }
 }
 
 bool shiro::replays::has_replay(const shiro::scores::score &s) {
-    return fs::exists(dir + fs::path::preferred_separator + std::to_string(s.id) + ".osr");
+    if (fs::exists(dir + fs::path::preferred_separator + std::to_string(s.id) + ".osr"))
+        return true;
+
+    return fs::exists(dir + fs::path::preferred_separator + std::to_string(s.id) + ".osr.zz");
 }
